csv_fill_data_gaps: Fixed a bug that would not fill the missing timestamp if it was at the very beginning or the very end of the dataset
test_csv_fill_data_gaps: Made a change to the original 4 datasets to work
with the new way the method works with start and end timestamps in the
middle of the week. Also added two mock files to be used in test that
require a weeks worth of data. I did this since it was a lot easier than
trying to create a dataframe in the test class with all the values
hardcoded.